backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
pydantic
python-dotenv
